+++ src/dipdup/models.py
@@ -140,10 +140,13 @@ class BlockData:
 class HeadBlockData:
     """Basic structure for head block from TzKT SignalR response"""
 
+    chain: str
+    chain_id: str
     cycle: int
     level: int
     hash: str
     protocol: str
+    next_protocol: str
     timestamp: datetime
     voting_epoch: int
     voting_period: int
@@ -158,6 +161,7 @@ class HeadBlockData:
     quote_jpy: Decimal
     quote_krw: Decimal
     quote_eth: Decimal
+    quote_gbp: Decimal
 
 
 @dataclass
@@ -240,7 +244,8 @@ class Contract(Model):
 
 
 class ContractMetadata(Model):
-    contract = fields.CharField(36, pk=True)
+    network = fields.CharField(51)
+    contract = fields.CharField(36)
     metadata = fields.JSONField()
     update_id = fields.IntField()
 
@@ -249,9 +254,11 @@ class ContractMetadata(Model):
 
     class Meta:
         table = 'dipdup_contract_metadata'
+        unique_together = ('network', 'contract')
 
 
 class TokenMetadata(Model):
+    network = fields.CharField(51)
     contract = fields.CharField(36)
     token_id = fields.IntField()
     metadata = fields.JSONField()
@@ -262,4 +269,4 @@ class TokenMetadata(Model):
 
     class Meta:
         table = 'dipdup_token_metadata'
-        unique_together = ('contract', 'token_id')
+        unique_together = ('network', 'contract', 'token_id')